implement change info page and enable personal info show

+++ Client/Dashboard/views.py
@@ -1,4 +1,4 @@
-from django.shortcuts import render
+from django.shortcuts import render,redirect
 from Dashboard import forms
 from Data import models
 import hashlib
@@ -10,10 +10,19 @@ def hash_code(s,salt='DataWrangler'):
 
 # Create your views here.
 def main(request):
-    return render(request,'main.html')
+    if not request.session.get('is_login',None):
+        return redirect('/login')
+    user = models.User.objects.get(user_id=request.session.get('user_id',None))
+    return render(request,'main.html',locals())
 def personal_info(request):
-    return render(request,'personal_info.html')
+    if not request.session.get('is_login',None):
+        return redirect('/login')
+    user = models.User.objects.get(user_id=request.session.get('user_id',None))
+    return render(request,'personal_info.html',locals())
 def change_password(request):
+    if not request.session.get('is_login',None):
+        return redirect('/login')
+    user = models.User.objects.get(user_id=request.session.get('user_id',None))
     if request.method == 'POST':
         change_password_form = forms.ChangePasswordForm(request.POST)
         message = "请检查填写的内容！"
@@ -21,11 +30,6 @@ def change_password(request):
             old_password = change_password_form.cleaned_data.get('old_password')
             new_password = change_password_form.cleaned_data.get('new_password')
             repeat_password = change_password_form.cleaned_data.get('repeat_password')
-            try:
-                user = models.User.objects.get(nickname=request.session.get('nickname',None))
-            except:
-                message = "用户不存在！"
-                return render(request,'change_password.html',locals())
             if user.password == hash_code(old_password):
                 if repeat_password == new_password:
                     user.password = new_password
@@ -40,7 +44,48 @@ def change_password(request):
                 return render(request,'change_password.html',locals())
     change_password_form = forms.ChangePasswordForm() 
     return render(request,'change_password.html',locals())
+def change_info(request):
+    if not request.session.get('is_login',None):
+        return redirect('/login')
+    user = models.User.objects.get(user_id=request.session.get('user_id',None))
+    if request.method == 'POST':
+        change_info_form = forms.ChangeInfoForm(request.POST)
+        message = "请检查填写的内容！"
+        if change_info_form.is_valid():
+            new_nickname = change_info_form.cleaned_data.get('nickname')
+            new_birth_date = change_info_form.cleaned_data.get('birth_date')
+            new_signature = change_info_form.cleaned_data.get('signature')
+            new_gender = change_info_form.cleaned_data.get('gender')
+            new_real_name = change_info_form.cleaned_data.get('real_name')
+            message = "成功！"
+            if(new_nickname != ""):
+                if models.User.objects.get(nickname=new_nickname):
+                    message = "用户名已被他人使用！"
+                    return render(request,'change_info.html',locals())
+                else:
+                    user.nickname=new_nickname
+                    user.save()
+                    request.session['nickname'] = new_nickname
+            if(new_birth_date != None):
+                user.birth_date = new_birth_date
+                user.save()
+            if(new_signature != ""):
+                user.signature = new_signature
+                user.save()
+            if(new_gender != ""):
+                user.gender = new_gender
+                user.save()
+            if(new_real_name != ""):
+                user.real_name = new_real_name
+                user.save()
+            return render(request,'change_info.html',locals())
+    message = "只需修改想修改的字段，其余留空。"
+    change_info_form = forms.ChangeInfoForm()
+    return render(request,'change_info.html',locals())
 def possession_show(request):
+    if not request.session.get('is_login',None):
+        return redirect('/login')
+    user = models.User.objects.get(user_id=request.session.get('user_id',None))
     try:
         devices = models.Device.objects.filter()
         return render(request,'possession_show.html',locals())
@@ -49,6 +94,9 @@ def possession_show(request):
         return render(request,'possession_show.html',locals())
     return render(request,'possession_show.html',locals())
 def possession_add(request):
+    if not request.session.get('is_login',None):
+        return redirect('/login')
+    user = models.User.objects.get(user_id=request.session.get('user_id',None))
     if request.method == 'POST':
         possession_add_form = forms.PossessionAddForm(request.POST)
         message = "请检查填写的内容！"
@@ -85,6 +133,9 @@ def possession_add(request):
     possession_add_form = forms.PossessionAddForm()
     return render(request,'possession_add.html',locals())
 def possession_delete(request):
+    if not request.session.get('is_login',None):
+        return redirect('/login')
+    user = models.User.objects.get(user_id=request.session.get('user_id',None))
     array = request.POST.getlist('checkbox')
     for id in array:
         models.Device.objects.get(device_id=id).delete()
